Removed demo extension and added own extension
