+++ src/main/resources/META-INF/services/qupath.lib.gui.extensions.QuPathExtension
@@ -1,2 +1,1 @@
-qupath.ext.template.DemoExtension
-qupath.ext.template.DemoGroovyExtension
+qupath.ext.gdcnn.GDCnnExtension
